+++ flumine/markets/markets.py
@@ -49,12 +49,12 @@ class Markets:
     @property
     def live_orders(self) -> bool:
         for market in self:
-            if market.closed is False and market.blotter.live_orders is True:
+            if market.closed is False and market.blotter.has_live_orders:
                 return True
         return False
 
     def __iter__(self) -> Iterator[Market]:
-        return iter(self.markets.values())
+        return iter(list(self.markets.values()))
 
     def __len__(self) -> int:
         return len(self.markets)