+++ flumine/streams/basestream.py
@@ -21,6 +21,8 @@ class BaseStream(threading.Thread):
         market_filter: dict = None,
         market_data_filter: dict = None,
         client=None,
+        output_queue: bool = True,
+        **listener_kwargs,
     ):
         threading.Thread.__init__(self, daemon=True, name=self.__class__.__name__)
         self.flumine = flumine
@@ -31,9 +33,11 @@ class BaseStream(threading.Thread):
         self.conflate_ms = conflate_ms
         self._client = client
         self._stream = None
-        self._output_queue = queue.Queue()
+        self._output_queue = queue.Queue() if output_queue else None
         self._listener = self.LISTENER(
-            output_queue=self._output_queue, max_latency=self.MAX_LATENCY
+            output_queue=self._output_queue,
+            max_latency=self.MAX_LATENCY,
+            **listener_kwargs,
         )
         self._output_thread = threading.Thread(
             name="{0}_output_thread".format(self._name),