+++ tests/test_baseflumine.py
@@ -81,7 +81,8 @@ class BaseFlumineTest(unittest.TestCase):
     def test__process_market_orders(self):
         mock_market = mock.Mock()
         mock_market.blotter.process_orders.return_value = [1, 2, 3]
-        self.base_flumine._process_market_orders(mock_market)
+        self.base_flumine.markets = [mock_market]
+        self.base_flumine._process_market_orders()
         mock_market.blotter.process_orders.assert_called_with(self.mock_client)
 
     def test__process_order_package(self):
@@ -168,7 +169,7 @@ class BaseFlumineTest(unittest.TestCase):
         mock_event = mock.Mock()
         self.base_flumine._process_custom_event(mock_event)
         mock_event.callback.assert_called_with(self.base_flumine, mock_event)
-        mock__process_market_orders.assert_called_with(mock_market)
+        mock__process_market_orders.assert_called_with()
 
     @mock.patch("flumine.baseflumine.BaseFlumine.info")
     @mock.patch("flumine.baseflumine.BaseFlumine.log_control")